Listen to fetched results controller's delegate methods
NSFetchedResultsController registers to listen change
notifications on managed object context and properly
updates its result set. After that, it notifies its delegate
about the changes (creation, update and deletion) of
underlying Core Data model (in this case Feed) via
methods defined in NSFetchedResultsControllerDelegate
protocol. We insert, delete or move rows in news table
view in response to those changes.

Note: It is important to send beginUpdates and endUpdates
messages to the table view instead of reloadData during this
update process.